moving the actions folder to src & other fixes
