--- dev/app/(frontend)/design-system/page.d.ts
@@ -1,1 +1,0 @@
-export default function Page(): Promise<import("react").JSX.Element>;
