+++ dev/app/(frontend)/design-system/page.tsx
@@ -1,5 +1,4 @@
-import { getTheme } from '@/actions/theme.actions'
-import { DesignSystemPage, LivePreviewListener } from 'theme-global/client'
+import { DesignSystemPage, getTheme, LivePreviewListener } from 'theme-global/client'
 
 export default async function Page() {
   const themeData = await getTheme()
