+++ dev/app/(frontend)/layout.tsx
@@ -1,24 +1,22 @@
-import { getTheme } from 'app/actions/theme.actions'
+import { FontHead, getTheme } from 'theme-global/client'
 
-import { FontHead } from '../../../src/components/theme/font-head'
 import './global.css'
 
 export default async function RootLayout({ children }: { children: React.ReactNode }) {
   const { cssVariables, fontCSS, themeData } = await getTheme()
+  const { fontBody, fontHeading } = themeData.typography || {}
 
   return (
     <html
       className="font-family"
-      dir={(cssVariables as any)['--theme-font-direction'] as string}
-      lang="en"
+      dir={(cssVariables as Record<string, string>)['--theme-font-direction'] || 'ltr'}
+      lang="en" 
       style={cssVariables}
       suppressHydrationWarning
     >
       <head>
-        {themeData.typography?.fontBody && <FontHead fontName={themeData.typography.fontBody} />}
-        {themeData.typography?.fontHeading && (
-          <FontHead fontName={themeData.typography.fontHeading} />
-        )}
+        {fontBody && <FontHead fontName={fontBody} />}
+        {fontHeading && <FontHead fontName={fontHeading} />}
         {fontCSS && <style dangerouslySetInnerHTML={{ __html: fontCSS }} />}
       </head>
       <body>{children}</body>
