+++ src/actions/client-theme.actions.d.ts
@@ -1,3 +1,3 @@
 export declare function getClientTheme({ noCache }?: {
     noCache?: boolean | undefined;
-}): Promise<import("../../../src/types").ThemeConfig>;
+}): Promise<import("../types").ThemeConfig>;
