+++ src/actions/theme.actions.d.ts
@@ -1,5 +1,5 @@
 import type * as React from 'react';
-import type { ThemeConfig } from '../../../src/types';
+import type { ThemeConfig } from '../types';
 export declare function getTheme({ noCache }?: {
     noCache?: boolean | undefined;
 }): Promise<{
