+++ src/exports/client.ts
@@ -1,2 +1,4 @@
-export { default as DesignSystemPage } from '../components/design-system/page.js'
-export { LivePreviewListener } from '../components/live-preview-listener/index.js'
+export { getTheme } from '../actions/theme.actions'
+export { default as DesignSystemPage } from '../components/design-system/page'
+export { LivePreviewListener } from '../components/live-preview-listener/index'
+export { FontHead } from '../components/theme/font-head'
